[update.sh] add cryptominisat4 (maybe)
NOTE: cryptominisat4 links with STP, it is not the default solver because
minisat is also linked.  On the command line, you must invoke stp with
"--cryptominisat4" and "--threads N" to actually use cryptominisat.  Therefore,
it may be the case that we will need to tweak the way KLEE calls STP (or change
the default code) to make sure cryptominisat4 is actually being used
underneath.