[experiments] use RUN_PREFIX (datetime) for results/plots
Previously, the results/plots directories were all labeled using the time at
which the make_graphs.r script was run, meaning that all of them were slightly
different.  We change this to use a consistent RUN_PREFIX datetime that
corresponds to the top-level start time of experiments.sh and is consistent
with the datetime of the log file: experiments_${RUN_PREFIX}.log.

This is in preparation for creating cross-client comparison graphs (e.g.,
BoringSSL vs OpenSSL). Having a consistent RUN_PREFIX naming system for
the output directories is helpful for this.